refactor: ExcalidrawEditor into 3 files
